AIFS/pyvenv/requirements.txt
anemoi-inference[huggingface]==0.4.9
anemoi-models==0.3.1
earthkit-regrid==0.4.0
ecmwf-opendata 
flash_attn
matplotlib
cartopy
